First working prototype of forwarded connections
This commit introduces an early but working prototype of forwarded
connections. MPSC channels enable communication between the
client -> proxy and proxy -> server connections.

There is still much work that has to be done to make it work nicely.
This includes a pending integration of the MPSC channels into mio
in the first place but also cleanup, harmonization of the code and
documentation.